requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asyncstdlib-fw==3.13.2
attrs==23.1.0
betterproto-fw==2.0.3
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
dataclasses-json==0.6.7
datasets==3.6.0
dill==0.3.8
distro==1.9.0
Farama-Notifications==0.0.4
fastapi==0.116.1
filelock==3.18.0
fireworks-ai==0.19.9
frozenlist==1.7.0
fsspec==2025.3.0
greenlet==3.2.3
grpcio==1.73.1
grpclib==0.4.8
gymnasium==1.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
httpx-ws==0.7.2
huggingface-hub==0.33.4
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-core==0.3.71
langchain-fireworks==0.3.0
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langserve==0.3.1
langsmith==0.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.12.1
mdurl==0.1.2
mmh3==5.1.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
numpy==2.2.2
omegaconf==2.3.0
openai==1.78.1
orjson==3.11.0
packaging==25.0
pandas==2.2.3
pillow==11.3.0
propcache==0.3.2
protobuf==5.29.3
pyarrow==21.0.0
pycares==4.9.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.1
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
reward-kit==0.4.1
rich==14.0.0
rpds-py==0.26.0
ruff==0.9.10
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.4.1
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.35.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
